24.12.2024_Memo_Call
Quick summary

During this meeting, Sergejs, Diāna, Nataļja, and Agita discussed the progress and challenges of their project, including the completion of tasks, the use of different tools for data analysis, and the preparation for the final presentation. They also addressed issues related to the organization of their GitHub repository and the need to update task descriptions to reflect the actual work done. The team also planned to finalize their project presentation by January 7th, with a follow-up meeting scheduled for December 29th to review the presentation details.